requirements/requirements.txt
# ===== 本番環境用依存関係 =====
# Discord.py関連
discord.py>=2.3.2
aiosqlite>=0.19.0
aiohttp>=3.8.5
aiofiles>=23.2.0

# Web関連（Keep-alive用）
flask>=2.3.3

# 音楽・音声関連
yt-dlp>=2023.7.6
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0

# 環境変数・設定
python-dotenv>=1.0.0

# HTTP関連
requests>=2.31.0

# JSON関連
json5>=0.9.14

# 音声合成（オプション）
# pyttsx3>=2.90